message crud

+++ base/views.py
@@ -1,15 +1,23 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
-from .models import Group, Topic
+from django.contrib.auth.forms import UserCreationForm
+from .models import Group, Topic, Message
 from .forms import GroupForm
 
 
 def login_page(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -24,14 +32,29 @@ def login_page(request):
             return redirect('home')
         else:
             messages.error(request, 'Username or password does not exist')
-
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 
 def logout_user(request):
     logout(request)
     return redirect('home')
+
+
+def register_page(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+    return render(request, 'base/login_register.html', {'form': form})
 
 
 def home(request):
@@ -53,12 +76,24 @@ def home(request):
 
 def group(request, pk):
     group = Group.objects.get(id=pk)
+    room_messages = group.message_set.all().order_by('-created')  # type: ignore
+    participants = group.participants.all()
 
-    context = {'group': group}
+    if request.method == 'POST':
+        Message.objects.create(
+            user=request.user,
+            group=group,
+            body=request.POST.get('body')
+        )
+        group.participants.add(request.user)
+        return redirect('group', pk=group.id)  # type: ignore
+
+    context = {'group': group, 'room_messages': room_messages, 'participants': participants}
 
     return render(request, 'base/group.html', context)
 
 
+@login_required(login_url='/login')
 def create_group(request):
     form = GroupForm()
     if request.method == 'POST':
@@ -71,9 +106,13 @@ def create_group(request):
     return render(request, 'base/group_form.html', context)
 
 
+@login_required(login_url='/login')
 def update_group(request, pk):
     group = Group.objects.get(id=pk)
     form = GroupForm(instance=group)
+
+    if request.user != group.host:
+        return HttpResponse('You are not allowed here')
 
     if request.method == 'POST':
         form = GroupForm(request.POST, instance=group)
@@ -85,10 +124,28 @@ def update_group(request, pk):
     return render(request, 'base/group_form.html', context)
 
 
+@login_required(login_url='/login')
 def delete_group(request, pk):
     group = Group.objects.get(id=pk)
+
+    if request.user != group.host:
+        return HttpResponse('You are not allowed here')
+
     if request.method == 'POST':
         group.delete()
         return redirect('home')
 
     return render(request, 'base/delete.html', {'obj': group})
+
+
+@login_required(login_url='/login')
+def delete_message(request, pk):
+    message = Message.objects.get(id=pk)
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+
+    return render(request, 'base/delete.html', {'obj': message})
